Update to SDL3 APIs and fix color depths
Implemention of SDL3 (if CMake definition ENABLE_SDL3 is set).
Implemention for setting of the color depth (both wxWidgets and SDL port).
Fix the coloring issues in the SDL port (OpenGL).
Implemention of 8 bit color support.
Implementation of SDL audio for wxWidgets port.
Implementation of SDL video for wxWidgets port.